Move title and bootstrap related helpers into their own files
